# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Material Library API",
    "description": "\n## Material Library Backend API\n\nA comprehensive API for managing materials, projects, products, components, and inventories \nwith environmental impact tracking (Global Warming Potential - GWP).\n\n### Features\n- **Material Database**: Create and manage materials with metadata and GWP values\n- **Project Management**: Organize projects with products and components\n- **Inventory Management**: Track material inventories with geolocation\n- **Material Transfer**: Transfer or gift materials between users\n- **User Relationships**: Connect with other users for material sharing\n\n### Authentication\nAll endpoints except registration and login require a valid JWT token.\nInclude the token in the Authorization header: `Bearer <token>`\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Material Library Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/api/materials/": {
      "get": {
        "tags": [
          "Materials"
        ],
        "summary": "Get all materials",
        "description": "Retrieve all materials with optional filtering",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in name and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "my_materials",
            "in": "query",
            "description": "Only show materials created by current user",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of results",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset for pagination",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of materials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Material"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Materials"
        ],
        "summary": "Create a new material",
        "description": "Create a new material with metadata and GWP values",
        "responses": {
          "201": {
            "description": "Material created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Recycled Steel"
                  },
                  "category": {
                    "type": "string",
                    "example": "Metals"
                  },
                  "description": {
                    "type": "string",
                    "example": "High-quality recycled steel"
                  },
                  "unit": {
                    "type": "string",
                    "example": "kg",
                    "default": "kg"
                  },
                  "gwp_value": {
                    "type": "number",
                    "example": 1.46,
                    "description": "Global Warming Potential per unit"
                  },
                  "gwp_unit": {
                    "type": "string",
                    "example": "kg CO2e",
                    "default": "kg CO2e"
                  },
                  "gwp_source": {
                    "type": "string",
                    "example": "EPD Database"
                  },
                  "is_reusable": {
                    "type": "boolean",
                    "default": false
                  },
                  "is_transferable": {
                    "type": "boolean",
                    "default": false
                  },
                  "is_giftable": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get user projects",
        "description": "Retrieve all projects owned by the current user",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status (active, archived, completed)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in name and description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create a new project",
        "description": "Create a new project",
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Office Building Renovation"
                  },
                  "description": {
                    "type": "string",
                    "example": "Sustainable renovation project"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "archived",
                      "completed"
                    ],
                    "default": "active"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/inventory/": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Get user inventory",
        "description": "Retrieve all inventory entries for the current user",
        "parameters": [
          {
            "name": "material_id",
            "in": "query",
            "description": "Filter by material ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_available",
            "in": "query",
            "description": "Filter by availability",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "available_for_transfer",
            "in": "query",
            "description": "Filter by transfer availability",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "available_for_gift",
            "in": "query",
            "description": "Filter by gift availability",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of inventory entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Inventory"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Create inventory entry",
        "description": "Add a material to user inventory with quantity and location",
        "responses": {
          "201": {
            "description": "Inventory entry created successfully"
          },
          "404": {
            "description": "Material not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "material_id",
                  "quantity"
                ],
                "properties": {
                  "material_id": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number",
                    "example": 100
                  },
                  "unit": {
                    "type": "string",
                    "example": "kg"
                  },
                  "location_name": {
                    "type": "string",
                    "example": "Main Warehouse"
                  },
                  "latitude": {
                    "type": "number",
                    "example": 52.52
                  },
                  "longitude": {
                    "type": "number",
                    "example": 13.405
                  },
                  "address": {
                    "type": "string",
                    "example": "Berlin, Germany"
                  },
                  "is_available": {
                    "type": "boolean",
                    "default": true
                  },
                  "available_for_transfer": {
                    "type": "boolean",
                    "default": false
                  },
                  "available_for_gift": {
                    "type": "boolean",
                    "default": false
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/relationships/": {
      "get": {
        "tags": [
          "Relationships"
        ],
        "summary": "Get user relationships",
        "description": "Retrieve all relationships for the current user",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status (pending, accepted, rejected)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "relationship_type",
            "in": "query",
            "description": "Filter by type (sharing, transfer)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of relationships"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Relationships"
        ],
        "summary": "Create relationship request",
        "description": "Send a relationship request to another user by email",
        "responses": {
          "201": {
            "description": "Relationship request created"
          },
          "400": {
            "description": "Relationship already exists or invalid request"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "receiver_email"
                ],
                "properties": {
                  "receiver_email": {
                    "type": "string",
                    "format": "email",
                    "example": "other@example.com"
                  },
                  "relationship_type": {
                    "type": "string",
                    "enum": [
                      "sharing",
                      "transfer"
                    ],
                    "default": "sharing"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/messages/": {
      "post": {
        "tags": [
          "Messages"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "receiver_id": {
                    "example": "any"
                  },
                  "inventory_id": {
                    "example": "any"
                  },
                  "subject": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Material": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "category": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "description": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "unit": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "default": {
                    "type": "string",
                    "example": "kg"
                  }
                }
              },
              "gwp_value": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  },
                  "description": {
                    "type": "string",
                    "example": "Global Warming Potential per unit"
                  }
                }
              },
              "gwp_unit": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "default": {
                    "type": "string",
                    "example": "kg CO2e"
                  }
                }
              },
              "gwp_source": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "is_reusable": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "is_transferable": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "is_giftable": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "created_by": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "created_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updated_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "description": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "status": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "enum": {
                    "type": "array",
                    "example": [
                      "active",
                      "archived",
                      "completed"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "owner_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "created_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updated_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      },
...
      "Inventory": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "user_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "material_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "uuid"
                  }
                }
              },
              "quantity": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "unit": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "location_name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "latitude": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "longitude": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "address": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "is_available": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "available_for_transfer": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "available_for_gift": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "boolean"
                  }
                }
              },
              "notes": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "created_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updated_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }
}